Split retained shell text input bridge
