+++ src/gui_runtime/native_shell_runtime/bridge.rs
@@ -1,7 +1,10 @@
+use self::pan_drag::WaveformPanDrag;
 use self::text_edit::{RetainedTextEditState, RetainedTextInputTarget};
 use super::*;
 
+mod pan_drag;
 mod text_edit;
+mod text_input;
 
 /// Wavecrate-owned generic Radiant runtime bridge.
 ///
@@ -35,13 +38,6 @@ pub(super) enum WavecrateRuntimeMessage {
     RetainedInput(WidgetInput),
     /// Local retained text-edit state changed without reducing an app action.
     LocalTextEdit,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-struct WaveformPanDrag {
-    anchor_x: f32,
-    start_micros: u32,
-    end_micros: u32,
 }
 
 impl<B> WavecrateRuntimeBridge<B> {
@@ -266,389 +262,6 @@ impl<B: NativeAppBridge> WavecrateRuntimeBridge<B> {
             .unwrap_or_else(|| Vector2::new(1280.0, 720.0));
         let style = StyleTokens::for_viewport_with_scale(viewport.x, 1.0);
         ShellLayout::build_with_style_and_runtime(viewport, &style, &mut self.layout_runtime)
-    }
-
-    fn begin_waveform_pan_drag(&mut self, layout: &ShellLayout, position: Point) {
-        self.waveform_pan_drag = None;
-        if !layout.waveform_plot.contains(position) {
-            return;
-        }
-        let viewport = self.model.waveform.viewport();
-        if viewport.end_micros.saturating_sub(viewport.start_micros) >= 999_999 {
-            return;
-        }
-        self.waveform_pan_drag = Some(WaveformPanDrag {
-            anchor_x: position.x,
-            start_micros: viewport.start_micros,
-            end_micros: viewport.end_micros,
-        });
-    }
-
-    fn waveform_pan_drag_action(&self, layout: &ShellLayout, position: Point) -> Option<UiAction> {
-        let drag = self.waveform_pan_drag?;
-        let span = drag.end_micros.saturating_sub(drag.start_micros).max(1);
-        if span >= 999_999 || layout.waveform_plot.width() <= 1.0 {
-            return None;
-        }
-        let delta_ratio = (position.x - drag.anchor_x) / layout.waveform_plot.width().max(1.0);
-        let delta_micros = (delta_ratio * span as f32).round() as i64;
-        let max_start = 1_000_000i64.saturating_sub(i64::from(span));
-        let start = (i64::from(drag.start_micros) - delta_micros).clamp(0, max_start);
-        let center_micros = (start + i64::from(span / 2)).clamp(0, 1_000_000) as u32;
-        Some(UiAction::SetWaveformViewCenter {
-            center_micros,
-            center_nanos: None,
-        })
-    }
-
-    /// Handle one non-text key intent routed through the focused retained canvas.
-    fn handle_retained_key_press(&mut self, key: WidgetKey) -> bool {
-        match key {
-            WidgetKey::Enter => {
-                if self.model.confirm_prompt.visible {
-                    self.emit_action(UiAction::ConfirmPrompt);
-                    return true;
-                }
-                if self.text_input_target == RetainedTextInputTarget::BrowserPillEditor {
-                    self.emit_action(UiAction::CommitBrowserTagSidebarInput);
-                }
-                self.text_input_target = RetainedTextInputTarget::None;
-                true
-            }
-            WidgetKey::Backspace => {
-                self.sync_text_edit_from_model();
-                if self.text_input_target == RetainedTextInputTarget::BrowserPillEditor
-                    && self.text_edit.value.is_empty()
-                    && !self.text_edit.has_selection()
-                {
-                    return self.remove_last_browser_pill_editor_chip();
-                }
-                self.edit_retained_text(|edit| edit.backspace())
-            }
-            WidgetKey::Delete => self.handle_retained_delete_key(),
-            WidgetKey::ArrowLeft => self.edit_retained_text(|edit| {
-                edit.move_caret(edit.caret.saturating_sub(1), false);
-                false
-            }),
-            WidgetKey::ArrowRight => self.edit_retained_text(|edit| {
-                edit.move_caret(edit.caret + 1, false);
-                false
-            }),
-            WidgetKey::Home => self.edit_retained_text(|edit| {
-                edit.move_caret(0, false);
-                false
-            }),
-            WidgetKey::End => self.edit_retained_text(|edit| {
-                edit.move_caret(edit.value.chars().count(), false);
-                false
-            }),
-            _ => false,
-        }
-    }
-
-    /// Handle one backend-level text edit command from Radiant's native runtime.
-    fn handle_retained_text_edit(&mut self, command: TextEditCommand) -> bool {
-        match command {
-            TextEditCommand::MoveLeft { extend_selection } => self.edit_retained_text(|edit| {
-                edit.move_caret(edit.caret.saturating_sub(1), extend_selection);
-                false
-            }),
-            TextEditCommand::MoveRight { extend_selection } => self.edit_retained_text(|edit| {
-                edit.move_caret(edit.caret + 1, extend_selection);
-                false
-            }),
-            TextEditCommand::MoveWordLeft { extend_selection } => self.edit_retained_text(|edit| {
-                edit.move_word_left(extend_selection);
-                false
-            }),
-            TextEditCommand::MoveWordRight { extend_selection } => {
-                self.edit_retained_text(|edit| {
-                    edit.move_word_right(extend_selection);
-                    false
-                })
-            }
-            TextEditCommand::MoveHome { extend_selection } => self.edit_retained_text(|edit| {
-                edit.move_caret(0, extend_selection);
-                false
-            }),
-            TextEditCommand::MoveEnd { extend_selection } => self.edit_retained_text(|edit| {
-                edit.move_caret(edit.value.chars().count(), extend_selection);
-                false
-            }),
-            TextEditCommand::SelectAll => self.edit_retained_text(|edit| {
-                edit.caret = edit.value.chars().count();
-                edit.selection_anchor = Some(0);
-                false
-            }),
-            TextEditCommand::InsertText(text) => {
-                let mut handled = false;
-                for character in text.chars() {
-                    if self.handle_retained_character(character) {
-                        handled = true;
-                    }
-                }
-                handled
-            }
-            TextEditCommand::Backspace => self.handle_retained_key_press(WidgetKey::Backspace),
-            TextEditCommand::Delete => self.handle_retained_key_press(WidgetKey::Delete),
-            TextEditCommand::DeleteWordLeft => {
-                self.edit_retained_text(RetainedTextEditState::delete_word_left)
-            }
-            TextEditCommand::DeleteWordRight => {
-                self.edit_retained_text(RetainedTextEditState::delete_word_right)
-            }
-            TextEditCommand::CutSelection => self.edit_retained_text(|edit| {
-                edit.replace_selection("");
-                true
-            }),
-        }
-    }
-
-    /// Insert one printable character into the active retained text target.
-    fn handle_retained_character(&mut self, character: char) -> bool {
-        if character.is_control() {
-            return false;
-        }
-        if self.model.confirm_prompt.visible && self.model.confirm_prompt.input_value.is_none() {
-            match character {
-                'y' | 'Y' => {
-                    self.emit_action(UiAction::ConfirmPrompt);
-                    return true;
-                }
-                'n' | 'N' => {
-                    self.emit_action(UiAction::CancelPrompt);
-                    return true;
-                }
-                _ => {}
-            }
-        }
-        self.sync_text_edit_from_model();
-        if self.text_input_target == RetainedTextInputTarget::BrowserPillEditor && character == ','
-        {
-            let token = self.text_edit.value.clone();
-            if token.split_whitespace().next().is_none() {
-                self.text_edit.value.clear();
-                self.text_edit.caret = 0;
-                self.text_edit.clear_selection();
-                return true;
-            }
-            self.emit_action(UiAction::SetBrowserTagSidebarInput {
-                value: token.clone(),
-            });
-            self.emit_action(UiAction::CommitBrowserTagSidebarInput);
-            self.text_input_target = RetainedTextInputTarget::BrowserPillEditor;
-            self.text_edit
-                .sync(RetainedTextInputTarget::BrowserPillEditor, String::new());
-            return true;
-        }
-        self.edit_retained_text(|edit| {
-            edit.insert_char(character);
-            true
-        })
-    }
-
-    fn handle_retained_delete_key(&mut self) -> bool {
-        if self.model.confirm_prompt.visible {
-            if self.text_input_target == RetainedTextInputTarget::Prompt
-                && self.model.confirm_prompt.input_value.is_some()
-            {
-                return self.edit_retained_text(|edit| edit.delete());
-            }
-            return true;
-        }
-        if self.text_input_target != RetainedTextInputTarget::None {
-            return self.edit_retained_text(|edit| edit.delete());
-        }
-        match self.model.focus_context {
-            runtime_contract::FocusContextModel::ContentList => {
-                self.emit_action(UiAction::DeleteBrowserSelection);
-                true
-            }
-            runtime_contract::FocusContextModel::NavigationTree => {
-                self.emit_action(UiAction::DeleteFocusedFolder);
-                true
-            }
-            _ => false,
-        }
-    }
-
-    /// Rewrite the active retained text target and emit the matching host action.
-    fn edit_retained_text(
-        &mut self,
-        rewrite: impl FnOnce(&mut RetainedTextEditState) -> bool,
-    ) -> bool {
-        if self.text_input_target == RetainedTextInputTarget::None {
-            return false;
-        }
-        self.sync_text_edit_from_model();
-        let changed = rewrite(&mut self.text_edit);
-        if !changed {
-            return true;
-        }
-        let value = self.text_edit.value.clone();
-        let action = match self.text_input_target {
-            RetainedTextInputTarget::None => return false,
-            RetainedTextInputTarget::BrowserSearch => UiAction::SetBrowserSearch { query: value },
-            RetainedTextInputTarget::FolderSearch => UiAction::SetFolderSearch { query: value },
-            RetainedTextInputTarget::BrowserPillEditor => {
-                UiAction::SetBrowserTagSidebarInput { value }
-            }
-            RetainedTextInputTarget::FolderCreate => UiAction::SetFolderCreateInput { value },
-            RetainedTextInputTarget::Prompt => UiAction::SetPromptInput { value },
-        };
-        self.emit_action(action);
-        true
-    }
-
-    fn sync_text_edit_from_model(&mut self) {
-        if let Some(value) = self.current_text_value() {
-            self.text_edit.sync(self.text_input_target, value);
-        }
-    }
-
-    fn remove_last_browser_pill_editor_chip(&mut self) -> bool {
-        let Some(pill) = self.model.browser.pill_editor().accepted_pills.last() else {
-            return true;
-        };
-        self.emit_action(UiAction::ToggleBrowserSidebarNormalTag {
-            label: pill.id.clone(),
-        });
-        true
-    }
-
-    /// Return the current projected text for the active retained text target.
-    fn current_text_value(&self) -> Option<String> {
-        match self.text_input_target {
-            RetainedTextInputTarget::None => None,
-            RetainedTextInputTarget::BrowserSearch => Some(self.model.browser.search_query.clone()),
-            RetainedTextInputTarget::FolderSearch => {
-                Some(self.model.sources.tree_search_query.clone())
-            }
-            RetainedTextInputTarget::BrowserPillEditor => {
-                Some(self.model.browser.pill_editor.input_value.clone())
-            }
-            RetainedTextInputTarget::FolderCreate => self.folder_inline_editor_value(),
-            RetainedTextInputTarget::Prompt => self.model.confirm_prompt.input_value.clone(),
-        }
-    }
-
-    fn folder_inline_editor_value(&self) -> Option<String> {
-        self.model
-            .sources
-            .tree_rows
-            .iter()
-            .find(|row| {
-                matches!(
-                    row.kind,
-                    runtime_contract::FolderRowKind::CreateDraft
-                        | runtime_contract::FolderRowKind::RenameDraft
-                )
-            })
-            .and_then(|row| row.input_value.clone())
-    }
-
-    /// Keep the local retained text target synchronized with host focus actions.
-    fn update_text_target_after_action(&mut self, action: &UiAction) {
-        self.text_input_target = match action {
-            UiAction::FocusBrowserSearch | UiAction::SetBrowserSearch { .. } => {
-                RetainedTextInputTarget::BrowserSearch
-            }
-            UiAction::FocusFolderSearch | UiAction::SetFolderSearch { .. } => {
-                RetainedTextInputTarget::FolderSearch
-            }
-            UiAction::FocusBrowserTagSidebarInput | UiAction::SetBrowserTagSidebarInput { .. } => {
-                RetainedTextInputTarget::BrowserPillEditor
-            }
-            UiAction::StartNewFolder
-            | UiAction::StartNewFolderAtFolderRow { .. }
-            | UiAction::StartNewFolderAtRoot
-            | UiAction::StartFolderRename
-            | UiAction::FocusFolderCreateInput
-            | UiAction::SetFolderCreateInput { .. } => RetainedTextInputTarget::FolderCreate,
-            UiAction::SetPromptInput { .. } => RetainedTextInputTarget::Prompt,
-            UiAction::BlurBrowserSearch
-            | UiAction::CommitBrowserTagSidebarInput
-            | UiAction::ConfirmFolderCreate
-            | UiAction::CancelFolderCreate
-            | UiAction::ConfirmPrompt
-            | UiAction::CancelPrompt
-            | UiAction::HandleEscape => RetainedTextInputTarget::None,
-            _ => self.text_input_target,
-        };
-        if self.text_input_target == RetainedTextInputTarget::FolderCreate
-            && self.folder_inline_editor_value().is_none()
-        {
-            self.text_input_target = RetainedTextInputTarget::None;
-        }
-        self.sync_text_edit_from_model();
-    }
-
-    fn resolve_retained_text_key_press(&mut self, press: RadiantKeyPress) -> bool {
-        if press.alt {
-            return false;
-        }
-        if press.command {
-            match press.key {
-                RadiantKeyCode::A => {
-                    self.text_edit.caret = self.text_edit.value.chars().count();
-                    self.text_edit.selection_anchor = Some(0);
-                    return true;
-                }
-                RadiantKeyCode::X => {
-                    self.text_edit.replace_selection("");
-                    let value = self.text_edit.value.clone();
-                    self.emit_text_value(value);
-                    return true;
-                }
-                RadiantKeyCode::C | RadiantKeyCode::V => return true,
-                _ => return false,
-            }
-        }
-        let len = self.text_edit.value.chars().count();
-        match press.key {
-            RadiantKeyCode::ArrowLeft => {
-                self.text_edit
-                    .move_caret(self.text_edit.caret.saturating_sub(1), press.shift);
-                true
-            }
-            RadiantKeyCode::ArrowRight => {
-                self.text_edit
-                    .move_caret((self.text_edit.caret + 1).min(len), press.shift);
-                true
-            }
-            RadiantKeyCode::Home => {
-                self.text_edit.move_caret(0, press.shift);
-                true
-            }
-            RadiantKeyCode::End => {
-                self.text_edit.move_caret(len, press.shift);
-                true
-            }
-            _ => false,
-        }
-    }
-
-    fn emit_text_value(&mut self, value: String) {
-        let action = match self.text_input_target {
-            RetainedTextInputTarget::None => return,
-            RetainedTextInputTarget::BrowserSearch => UiAction::SetBrowserSearch { query: value },
-            RetainedTextInputTarget::FolderSearch => UiAction::SetFolderSearch { query: value },
-            RetainedTextInputTarget::BrowserPillEditor => {
-                UiAction::SetBrowserTagSidebarInput { value }
-            }
-            RetainedTextInputTarget::FolderCreate => UiAction::SetFolderCreateInput { value },
-            RetainedTextInputTarget::Prompt => UiAction::SetPromptInput { value },
-        };
-        self.emit_action(action);
-    }
-
-    fn apply_local_text_projection(&self, model: &mut runtime_contract::AppModel) {
-        if self.text_input_target != RetainedTextInputTarget::BrowserPillEditor {
-            return;
-        }
-        model.browser.pill_editor.input_focused = true;
-        model.browser.pill_editor.input_caret = self.text_edit.caret;
-        model.browser.pill_editor.input_selection = self.text_edit.selection();
     }
 }
